refactor: Restructured project directory and improved Bull setup
Renamed 'models' folder to 'modules' for better project organization.
Introduced constants for SMTP queue name and configuration settings.
Created a 'queues' folder to house specific channel-related code.
Updated Bull module loading to use async configuration for Redis.
