+++ src/jobs/consumers/notifications/smtp-notifications.job.consumer.ts
@@ -3,13 +3,14 @@ import { Processor, Process } from '@nestjs/bull';
 import { Job } from 'bull';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Notification } from 'src/models/notifications/entities/notification.entity';
+import { Notification } from 'src/modules/notifications/entities/notification.entity';
 import { DeliveryStatus } from 'src/common/constants/notifications';
 import { SmtpService } from 'src/services/email/smtp/smtp.service';
 import * as nodemailer from 'nodemailer';
-import { NotificationsService } from 'src/models/notifications/notifications.service';
+import { NotificationsService } from 'src/modules/notifications/notifications.service';
+import { SMTP_QUEUE } from 'src/modules/notifications/queues/smtp.queue';
 
-@Processor('smtpNotifications')
+@Processor(SMTP_QUEUE)
 export class SmtpNotificationConsumer {
   private readonly logger = new Logger(SmtpNotificationConsumer.name);
 
@@ -21,7 +22,7 @@ export class SmtpNotificationConsumer {
   ) {}
 
   @Process()
-  async processSmtpNotificationQueue(job: Job): Promise<void> {
+  async processSmtpNotificationQueue(job: Job<number>): Promise<void> {
     const id = job.data;
     const notification = (await this.notificationsService.getNotificationById(id))[0];
