+++ src/jobs/producers/notifications/notifications.job.producer.ts
@@ -1,11 +1,12 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { InjectQueue } from '@nestjs/bull';
 import { Queue } from 'bull';
-import { Notification } from 'src/models/notifications/entities/notification.entity';
+import { Notification } from 'src/modules/notifications/entities/notification.entity';
+import { SMTP_QUEUE } from 'src/modules/notifications/queues/smtp.queue';
 
 @Injectable()
 export class NotificationQueueProducer {
-  constructor(@InjectQueue('smtpNotifications') private readonly smtpQueue: Queue) {}
+  constructor(@InjectQueue(SMTP_QUEUE) private readonly smtpQueue: Queue) {}
   private readonly logger = new Logger(NotificationQueueProducer.name);
   async onModuleInit(): Promise<void> {
     this.smtpQueue.client.on('error', (error) => {
