+++ src/modules/notifications/queues/smtp.queue.ts
@@ -1,0 +1,5 @@
+export const SMTP_QUEUE = 'smtp';
+
+export const smtpQueueConfig = {
+  name: SMTP_QUEUE,
+};
